Move the OpenDapHelper.testDapToNcDGrid to maven tests
Also comment out ByteArrayTests from TestAll. Maven can't compile that since it's production code relying on test code. The right solution is likely moving TestAll to be within the test sources, but I'll need to verify that has no side effects on other processes.

testDapToNcDGrid is currently failing, but I believe it's due to the data being returned from the server request.  I've added a todo in the code about plans for fixing that.